+++ bubble-api/bubble-api-oa/src/main/java/com/bubblecloud/oa/api/vo/frame/FrameDetailVO.java
@@ -19,9 +19,9 @@ public class FrameDetailVO {
 
 	private Long entid;
 
-	private Integer pid;
+	private Long pid;
 
-	private Integer roleId;
+	private Long roleId;
 
 	private String name;
